edit joystick lib

+++ lib/joystick/src/joystick.h
@@ -4,14 +4,14 @@
 class joystick
 {
 public:
-    float x;
-    float y;
-    float xmax;
-    float ymax;
-    float xmin;
-    float ymin;
-    joystick(float _x, float _y, float _xmax, float _ymax, float _xmin, float _ymin);
-    float vel;
-    float phi;
+    int x;
+    int y;
+    int xmax;
+    int ymax;
+    int xmin;
+    int ymin;
+    joystick(int _x, int _y, int _xmax, int _ymax, int _xmin, int _ymin);
+    int vel;
+    int phi;
 };
 #endif
